chore: reorganised serialise versions
