+++ src/model/serialise/project.ts
@@ -1,23 +1,10 @@
-import { compress, decompress } from '@helpers/compress';
-import {
-  dateToISOString,
-  formatShortDate,
-  getDateFromUnixTime,
-  getUnixTimeFromDate
-} from '@helpers/datetime';
+import { dateToISOString, formatShortDate } from '@helpers/datetime';
 import { createLog } from '@helpers/log';
 import { generateShortUUID } from '@helpers/uuid';
-import {
-  exportPadToJSON,
-  exportPadToURLString,
-  importPadFromJSON,
-  importPadFromURLString
-} from '@model/serialise/pad';
+import { exportPadToJSON, importPadFromJSON } from '@model/serialise/pad';
 import {
   exportSequencerToJSON,
-  exportSequencerToURLString,
-  importSequencerFromJSON,
-  importSequencerFromURLString
+  importSequencerFromJSON
 } from '@model/serialise/sequencer';
 import { addOrReplacePad } from '@model/store/actions/helpers';
 import { initialContext } from '@model/store/store';
@@ -25,12 +12,14 @@ import { ProjectStoreContext, ProjectStoreType } from '@model/store/types';
 import { PadExport, ProjectExport } from '@model/types';
 import {
   exportStepSequencerToJSON,
-  exportStepSequencerToURLString,
-  importStepSequencerFromJSON,
-  importStepSequencerFromURLString
+  importStepSequencerFromJSON
 } from './stepSequencer';
+import { exportToURLStringV1, importFromURLStringV1 } from './versions/1';
+import { exportToURLStringV2, importFromURLStringV2 } from './versions/2';
+import { exportToURLStringV3, importFromURLStringV3 } from './versions/3';
+import { exportToURLStringV4, importFromURLStringV4 } from './versions/4';
 
-const log = createLog('model/export', ['debug']);
+const log = createLog('serialise/project');
 
 const EXPORT_JSON_VERSION = '2025-02-18';
 const EXPORT_APP_VERSION = process.env.VERSION;
@@ -106,148 +95,6 @@ export const exportToURLString = async (
   }
 };
 
-const exportToURLCommon = (project: ProjectStoreType) => {
-  const { context } = project.getSnapshot();
-
-  const {
-    projectId,
-    projectName,
-    projectBgImage,
-    createdAt,
-    updatedAt,
-    pads,
-    sequencer
-  } = context;
-
-  const sequencerURL = sequencer ? exportSequencerToURLString(sequencer) : '';
-
-  const padsURL = pads.map((pad) => exportPadToURLString(pad)).filter(Boolean);
-
-  const createTimeSecs = getUnixTimeFromDate(createdAt);
-  const updateTimeSecs = getUnixTimeFromDate(updatedAt);
-
-  return {
-    context,
-    projectId,
-    projectName,
-    projectBgImage,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  };
-};
-
-export const exportToURLStringV1 = (project: ProjectStoreType) => {
-  const {
-    projectId,
-    projectName,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  } = exportToURLCommon(project);
-
-  // base64 encode the project name
-  const projectNameBase64 = projectName
-    ? btoa(encodeURIComponent(projectName))
-    : '';
-
-  let result = `1|${projectId}|${projectNameBase64}|${createTimeSecs}|${updateTimeSecs}`;
-
-  result = addPadsAndSequencerToResult(result, padsURL, sequencerURL);
-
-  return result;
-};
-
-const addPadsAndSequencerToResult = (
-  result: string,
-  padsURL: (string | undefined)[],
-  sequencerURL?: string
-) => {
-  if (padsURL.length > 0) {
-    result += `|${padsURL.join('(')}`;
-  } else {
-    result += '|';
-  }
-
-  if (sequencerURL) {
-    result += `|${sequencerURL}`;
-  }
-
-  return result;
-};
-
-export const exportToURLStringV2 = async (project: ProjectStoreType) => {
-  const {
-    projectId,
-    projectName,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  } = exportToURLCommon(project);
-
-  let result = `${projectId}|${projectName}|${createTimeSecs}|${updateTimeSecs}`;
-
-  result = addPadsAndSequencerToResult(result, padsURL, sequencerURL);
-
-  const compressed = await compress(result);
-
-  return `2|${compressed}`;
-};
-
-export const exportToURLStringV3 = async (project: ProjectStoreType) => {
-  const {
-    projectId,
-    projectName,
-    projectBgImage,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  } = exportToURLCommon(project);
-
-  let result = `${projectId}|${projectName}|${projectBgImage}|${createTimeSecs}|${updateTimeSecs}`;
-
-  result = addPadsAndSequencerToResult(result, padsURL, sequencerURL);
-
-  const compressed = await compress(result);
-
-  return `3|${compressed}`;
-};
-
-export const exportToURLStringV4 = async (project: ProjectStoreType) => {
-  const {
-    context,
-    projectId,
-    projectName,
-    projectBgImage,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  } = exportToURLCommon(project);
-
-  const { stepSequencer } = context;
-
-  const stepSequencerURL = stepSequencer
-    ? exportStepSequencerToURLString(stepSequencer)
-    : '';
-
-  let result = `${projectId}|${projectName}|${projectBgImage}|${createTimeSecs}|${updateTimeSecs}`;
-
-  result = addPadsAndSequencerToResult(result, padsURL, sequencerURL);
-
-  if (stepSequencerURL) {
-    result += `|${stepSequencerURL}`;
-  }
-
-  const compressed = await compress(result);
-
-  return `4|${compressed}`;
-};
-
 export const importProjectExport = (
   data: ProjectExport
 ): ProjectStoreContext => {
@@ -306,6 +153,8 @@ export const urlStringToProject = async (urlString: string) => {
   const version = urlString.slice(0, firstPipe);
   const data = urlString.slice(firstPipe + 1);
 
+  log.debug('urlStringToProject', urlString, version, data);
+
   if (version === '1') {
     return importFromURLStringV1(data);
   }
@@ -320,154 +169,4 @@ export const urlStringToProject = async (urlString: string) => {
   }
 
   throw new Error(`Unsupported export version: ${version}`);
-};
-
-const importFromURLStringV1 = (urlString: string) => {
-  log.debug('importFromURLStringV1', urlString);
-  const [
-    projectId,
-    projectNameBase64,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  ] = urlString.split('|');
-
-  const createdAt = getDateFromUnixTime(createTimeSecs);
-  const updatedAt = getDateFromUnixTime(updateTimeSecs);
-
-  const projectName = decodeURIComponent(atob(projectNameBase64));
-
-  const pads = padsURL
-    .split('(')
-    .map(importPadFromURLString)
-    .filter(Boolean) as PadExport[];
-
-  const sequencer = sequencerURL
-    ? importSequencerFromURLString(sequencerURL)
-    : undefined;
-
-  return {
-    id: projectId,
-    name: projectName,
-    createdAt: dateToISOString(createdAt),
-    updatedAt: dateToISOString(updatedAt),
-    pads,
-    sequencer
-  } as ProjectExport;
-};
-
-const importFromURLStringV2 = async (data: string) => {
-  const uncompressed = await decompress(data);
-
-  log.debug('importFromURLStringV2', uncompressed);
-  const [
-    projectId,
-    projectName,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  ] = uncompressed.split('|');
-
-  const createdAt = getDateFromUnixTime(createTimeSecs);
-  const updatedAt = getDateFromUnixTime(updateTimeSecs);
-
-  const pads = padsURL
-    .split('(')
-    .map(importPadFromURLString)
-    .filter(Boolean) as PadExport[];
-
-  const sequencer = sequencerURL
-    ? importSequencerFromURLString(sequencerURL)
-    : undefined;
-
-  return {
-    id: projectId,
-    name: projectName,
-    createdAt: dateToISOString(createdAt),
-    updatedAt: dateToISOString(updatedAt),
-    pads,
-    sequencer
-  } as ProjectExport;
-};
-
-const importFromURLStringV3 = async (data: string) => {
-  const uncompressed = await decompress(data);
-
-  log.debug('importFromURLStringV3', uncompressed);
-  const [
-    projectId,
-    projectName,
-    projectBgImage,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL
-  ] = uncompressed.split('|');
-
-  const createdAt = getDateFromUnixTime(createTimeSecs);
-  const updatedAt = getDateFromUnixTime(updateTimeSecs);
-
-  const pads = padsURL
-    .split('(')
-    .map(importPadFromURLString)
-    .filter(Boolean) as PadExport[];
-
-  const sequencer = sequencerURL
-    ? importSequencerFromURLString(sequencerURL)
-    : undefined;
-
-  return {
-    id: projectId,
-    name: projectName,
-    bgImage: projectBgImage,
-    createdAt: dateToISOString(createdAt),
-    updatedAt: dateToISOString(updatedAt),
-    pads,
-    sequencer
-  } as ProjectExport;
-};
-
-const importFromURLStringV4 = async (data: string) => {
-  const uncompressed = await decompress(data);
-
-  log.debug('importFromURLStringV4', uncompressed);
-  const [
-    projectId,
-    projectName,
-    projectBgImage,
-    createTimeSecs,
-    updateTimeSecs,
-    padsURL,
-    sequencerURL,
-    stepSequencerURL
-  ] = uncompressed.split('|');
-
-  const createdAt = getDateFromUnixTime(createTimeSecs);
-  const updatedAt = getDateFromUnixTime(updateTimeSecs);
-
-  const pads = padsURL
-    .split('(')
-    .map(importPadFromURLString)
-    .filter(Boolean) as PadExport[];
-
-  const sequencer = sequencerURL
-    ? importSequencerFromURLString(sequencerURL)
-    : undefined;
-
-  const stepSequencer = stepSequencerURL
-    ? importStepSequencerFromURLString(stepSequencerURL)
-    : undefined;
-
-  return {
-    id: projectId,
-    name: projectName,
-    bgImage: projectBgImage,
-    createdAt: dateToISOString(createdAt),
-    updatedAt: dateToISOString(updatedAt),
-    pads,
-    sequencer,
-    stepSequencer
-  } as ProjectExport;
 };
